support switching between cooked and raw mode
Stop using commitizen commit format because it doesn't make a lot
of sense for the project.

The solution I came up with for manipulating the terminal isn't very
pretty with this commit. For what it's worth, at least it is only
written in one place.

We call out to libc to set the terminal to "raw" mode to accept
single-byte keybindings.