fix(PieceFile): add action pieces in proper order
When undoing a delete we were adding pieces to the
action in the order they appear rather than in reverse.

Really we could just pop them off one at a time when
undoing, but this makes things a little bit cleaner.

In the future it might be better just to pop them off
instead of looping, but it doesn't honestly matter
that much. The performance gains would be negligible.