Add invertible undos and redos (#1)
* chore(PieceFile): small cleanups

* fix(PieceFile): refactor delete to be better

* feat(PieceFile): don't automatically add actions

In the future we'll need to do deletes and inserts
without caring about the actions that they generate.
We don't want them to pollute the undo stream.

This exposes `delete` and `insert` as wrappers around
the internal functions.

I'm not sure if this is idiomatic, but it works.

* feat(PieceFile): partially implement undo

* feat(PieceTable): handle merging pieces and undos

* feat(PieceFile): only merge pieces post-undo if necessary

* feat(PieceFile): finish implementing undo

Also include some tests. They're really more integration
than unit, but hey, they get the job done.

* fix(PieceFile): add action pieces in proper order

When undoing a delete we were adding pieces to the
action in the order they appear rather than in reverse.

Really we could just pop them off one at a time when
undoing, but this makes things a little bit cleaner.

In the future it might be better just to pop them off
instead of looping, but it doesn't honestly matter
that much. The performance gains would be negligible.

* feat(PieceFile): allow for redoing undos

* fix(PieceFile): handle upper-end deletes

There was a problem with the PieceFile wherein if a delete
occurred at the latter half of a piece and did not go into
the next, any subsequent undo would produce incorrect results.

This commit handles the edge case and adds tests that check for it.

* fix(PieceFile): switch improper use of variable

* fix(PieceFile): correct failing test

* fix(PieceFile): don't error on many undos

* feat(PieceFile): add timestamp to Actions

* feat(PieceFile): remove newer actions on change

If the user undoes a change and makes a new change,
the change the user undid is removed from the action
history. This is because the action history is stored
as a stack instead of a tree, as vim does it.

* feat(README): check boxes