feat(Keymaster): holy shit, it works
This rewrite makes it so that the Keymaster doesn't need to mess with
lifetimes. The problem with the previous approach was that it was
nontrivial to keep the lifetimes in sync and have the Keymaster just
make what it needed on the stack. Ideally in the future this stuff could
all be stored on the heap, but this approach works just fine for now.