Improve bindings and mutations of state (#5)
* fix(binds): cut out all of the unnecessary stuff

We're redoing the way bindings are interpreted because the former system
didn't allow the required level of customiziability.

* feat(Keymaster): remove EVERYTHING

* feat(Keymaster): holy shit, it works

This rewrite makes it so that the Keymaster doesn't need to mess with
lifetimes. The problem with the previous approach was that it was
nontrivial to keep the lifetimes in sync and have the Keymaster just
make what it needed on the stack. Ideally in the future this stuff could
all be stored on the heap, but this approach works just fine for now.

* feat(Keymaster): handle action updates

* feat(Keymaster): fix consume()

* feat(Keymaster): add Root arrow

* feat(Keymaster): try to make make_tables

* feat(binds): reintroduce tests, add make_table

* feat(binds): introduce other tests

* feat(Keymaster): rename method

* feat(Keymaster): get arrow from sequence

Add method that grabs an arrow given a sequence of
keys.

* fix(Arrow): change docstring

* feat(byt): add Editor for shared mutation

* feat(byt): change call

* feat(byt): use traits instead

This moves action consumption and key input to traits so that they can
be used a bit more easily in different contexts.

* feat(Editor): implement Renderable

* feat(Editor): get Editor rendering working

* feat(Keymaster): add bind method for sequences

* feat(Keymaster): sequence based unbinds

* feat(Keymaster): small changes

* feat(editor): add mutators

* feat(RustMutator): closure storage for Rust

The RustMutator stores closures that mutate a provided type.
The goal is that they can be used for editor and file local
modifications.

* feat(RustMutator): give state parameter too

* chore(RustScope): rename for clarity

* feat(RustScope): change to store state on its own

* feat(Editor): add Mutatable trait

* feat(MutablePair): use a mutation wrapper

* feat(MutatorPair): add struct for implicit mutation

* feat(Vym): implement basic vym skeleton

vym is the vim mode for byt

* feat(Vym): janky implementation working

* feat(Editor): polish up rendering order