feat(FileView): attempt to fix rendering issue
We used to redraw the entire screen on every action performed, but
that didn't strike me as a sustainable model. Now renders are split up
into line changes and cursor changes; the former requires a full redraw.
However, the flickering problem is still there. Eventually I will have
to reimplement the TermRenderer to only write characters if they're
different from what is already on the screen.