feat(byt): remove insertion code
Soooo the insertion strategy we were using before to batch writes to the
piece file wasn't all that great. It resulted in a lot more complexity
in the FileView that didn't need to be there. Ultimately it would be
better for the PieceFile to combine writes together that consist of one
byte apiece and happen sequentially.